Support coverage report per modules (#44)
* :fire: Remove unit test

This commit is to remove unused unit test.

* :wrench: update launch configuration

This commit is to update launch.json and settings.json.

* 🚧 Add MessageBuilder

This commit is to add MessageBuilder class
which set coverage per modules and generate output message.

* ✨ Update coverage report per file

This commit is to update coverage message which output per files.

* 🐛 Fix bug of failed to create comment

* Fix bug of output message

This commit is to fix the bug which output message per file.
It expect message per modules(ex. dll, module).